+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## 2.3.0 January 23, 2025
 
+- Enh #211: Add ability to terminate the processes that taken the port (@xepozz)
 - Enh #207: Add `--open` option for `serve` command (@xepozz)
 - Enh #207: Print possible options for `serve` command (@xepozz)
 - Enh #218: Explicitly mark nullable parameters (@alexander-schranz)